[NU-2469] refactors
