+++ designer/client/src/components/toolbars/assertionResults/assertionResultsForNode/assertionResult/AssertionStatusIcon.tsx
