+++ designer/client/src/components/toolbars/assertionResults/assertionResultsForNode/assertionResult/assertionWithDetails.tsx
