+++ designer/client/src/components/toolbars/assertionResults/assertionResultsForNode/assertionResultsForNode.tsx
@@ -1,13 +1,14 @@
 import { Box, Typography } from "@mui/material";
-import React, {Fragment} from "react";
+import React, { Fragment } from "react";
 
 import { getTestCaseAssertionsForNode } from "../../../../reducers/selectors/testCases";
 import { getTestAssertionResultsForNode } from "../../../../reducers/selectors/testing";
 import { useAppSelector } from "../../../../store/storeHelpers";
 import { InfoTooltip } from "../../../graph/node-modal/editors/InfoTooltip/InfoTooltip";
-import { AssertionExpression } from "./assertionExpression";
-import { AssertionResultMessage } from "./assertionResultMessage";
-import { AssertionStatusIcon } from "./AssertionStatusIcon";
+import { AssertionExpression } from "./assertionResult/assertionExpression";
+import { AssertionResult } from "./assertionResult/assertionResult";
+import { AssertionResultMessage } from "./assertionResult/assertionResultMessage";
+import { AssertionStatusIcon } from "./assertionResult/AssertionStatusIcon";
 
 interface Props {
     nodeId: string;
@@ -21,40 +22,12 @@ export const AssertionResultsForNode = ({ nodeId }: Props) => {
         <Box display={"flex"} flexDirection={"column"} gap={0.25}>
             {testCaseAssertions.map((testCaseAssertion, index) => {
                 const assertionResult = assertionResults?.[index];
+
                 if (!assertionResult) {
-                    return null;
+                    return;
                 }
-
-                const tooltipContent = (
-                    <Box p={1}>
-                        {testCaseAssertion.description && <AssertionExpression assertion={testCaseAssertion} />}
-                        <AssertionResultMessage assertionResult={assertionResult} />
-                    </Box>
-                );
-
-                const displayTooltip = assertionResult.type === "FailedAssertion" || testCaseAssertion.description;
-
-                const content = (
-                    <Box display={"flex"} gap={0.75} alignItems={"center"} sx={{ cursor: displayTooltip && "pointer" }}>
-                        <AssertionStatusIcon isSuccess={assertionResult.type === "SuccessfulAssertion"} />
-                        {testCaseAssertion.description ? (
-                            <Typography variant={"caption"}>{testCaseAssertion.description}</Typography>
-                        ) : (
-                            <AssertionExpression assertion={testCaseAssertion} />
-                        )}
-                    </Box>
-                );
-
                 return (
-                    <Fragment key={testCaseAssertion.uuid}>
-                        {displayTooltip ? (
-                            <InfoTooltip variant={"click"} title={tooltipContent} placement={"right"}>
-                                {content}
-                            </InfoTooltip>
-                        ) : (
-                            content
-                        )}
-                    </Fragment>
+                    <AssertionResult key={testCaseAssertion.uuid} assertionResult={assertionResult} testCaseAssertion={testCaseAssertion} />
                 );
             })}
         </Box>
